# server/Swagger/OpenAPI.yml
openapi: 3.0.1
info:
  title: SE-Sept-14 API
  description: API endpoints for Recommender System project of Group 14
  version: 0.1.0
...
paths:
  /api/auth/register:
    post:
      tags:
        - Auth
      summary: Create a new user
      operationId: register_user_api_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: Login User
      operationId: login_user_api_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    UserLogin:
      properties:
        email:
          type: string
          title: Email
        password:
          type: string
          title: Password
      type: object
      required:
        - email
        - password
      title: UserLogin
...
    UserRegistration:
      properties:
        email:
          type: string
          title: Email
        password:
          type: string
          title: Password
        is_admin:
          anyOf:
            - type: integer
            - nullable: true
          title: Is Admin
        profile:
          anyOf:
            - type: string
            - nullable: true
          title: Profile
        username:
          type: string
          title: Username
      type: object
      required:
        - email
        - password
        - username
      title: UserRegistration
    UserResponse:
      properties:
        id:
          type: integer
          title: Id
        email:
          type: string
          title: Email
        is_admin:
          type: integer
          title: Is Admin
        profile:
          type: string
          title: Profile
        username:
          type: string
          title: Username
      type: object
      required:
        - id
        - email
        - is_admin
        - profile
        - username
      title: UserResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/auth/login